+++ include/Towers/CheesePizza.hpp
@@ -19,10 +19,10 @@ public:
   void setUpUnits(){}
   void setPos(intPair pos){this->row=pos.first;this->col=pos.second;}
   void setPos(int row, int col){this->row=row; this->col=col;}
-  int getXCoordinate(){return xCoordinate;}
-  int getYCoordinate(){return yCoordinate;}
-  void setXCoordinate(int xCor){xCoordinate = xCor;}
-  void setYCoordinate(int yCor){yCoordinate = yCor;}
+  float getXCoordinate(){return xCoordinate;}
+  float getYCoordinate(){return yCoordinate;}
+  void setXCoordinate(float xCor){xCoordinate = xCor;}
+  void setYCoordinate(float yCor){yCoordinate = yCor;}
   bool canAttack(){return RangeTower::canAttack();}
   void attack(shared_ptr<ActorInterface> enemyInRange){RangeTower::attack(enemyInRange);}
   void update(float delta){RangeTower::update(delta);}